Create trainingDays.js
The main theme of this projects is using knowledge and understanding of variable scopes. 

As a seasoned athlete, one of your favorite activities is running marathons. You use a service called Training Days that sends you a message for the event you signed up for and the days you have left to train.